+++ include/executor.h
@@ -49,10 +49,10 @@ void						execute(t_cmdtable *table, t_dt *minishell,
 int							ft_open(char *file, t_token_type rdr_type);
 char						*ft_strgetbetween(const char *start,
 								const char *end);
-void							exe_parent_builtin_cmds(t_cmdtable *table,
+void						exe_parent_builtin_cmds(t_cmdtable *table,
 								t_dt *minishell);
 void						child(t_cmdtable *table, bool last_cmd,
-								t_env *envp_lst, int *fd);
+								t_dt *minishell, int *fd);
 
 // redirector.c
 int							check_redirections(t_cmdtable *table, int *fd);
@@ -69,12 +69,13 @@ void						unset(t_env **head, char **var);
 void						ft_waitpid(t_dt *minishell, int nb_cmd);
 char						*get_env(char **env, const char *name);
 char						*cmd_path(char *cmd, char **env);
+int							is_valid_env_name(const char *name);
 
 // exe_utils1.c
 void						free_arrayofstrings(char **array);
 t_env						*create_env_var_node(char *str);
 void						print_env_ascending(t_env *head);
-int	count_env_variables(t_env *head);
+int							count_env_variables(t_env *head);
 // signals.c
 void						block_signal(void);
