+++ src/executor/builtin_cmds/pwd.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   built_in_cmds.c                                    :+:      :+:    :+:   */
+/*   pwd.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ekordi <ekordi@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/01 15:27:46 by ekordi            #+#    #+#             */
-/*   Updated: 2024/01/01 15:30:49 by ekordi           ###   ########.fr       */
+/*   Updated: 2024/01/20 14:48:26 by ekordi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,4 +28,3 @@ int	execute_pwd(void)
 	free(pwd);
 	return (1);
 }
-
