+++ src/executor/utils/exe_utils2.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   exe_utils2.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kglebows <kglebows@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ekordi <ekordi@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/30 16:42:42 by ekordi            #+#    #+#             */
-/*   Updated: 2024/01/18 20:11:53 by kglebows         ###   ########.fr       */
+/*   Updated: 2024/01/20 14:56:50 by ekordi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,8 +23,8 @@ static void	assign_array(t_env *head, char **array)
 	{
 		if (temp->value)
 		{
-			array[i] = (char *)malloc(ft_strlen(temp->key) + ft_strlen(temp->value)
-				+ 2);
+			array[i] = (char *)malloc(ft_strlen(temp->key)
+					+ ft_strlen(temp->value) + 2);
 			ft_strlcpy(array[i], temp->key, ft_strlen(temp->key) + 1);
 			ft_strlcat(array[i], "=", ft_strlen(array[i]) + 2);
 			ft_strlcat(array[i], temp->value, ft_strlen(array[i])
@@ -35,6 +35,7 @@ static void	assign_array(t_env *head, char **array)
 	}
 	array[i] = NULL;
 }
+
 int	count_env_variables_with_value(t_env *head)
 {
 	int		i;
@@ -50,6 +51,7 @@ int	count_env_variables_with_value(t_env *head)
 	}
 	return (i);
 }
+
 char	**env_to_char_array(t_env *head)
 {
 	int		count;
@@ -59,62 +61,6 @@ char	**env_to_char_array(t_env *head)
 	result = (char **)calloc((count + 1), sizeof(char *));
 	assign_array(head, result);
 	return (result);
-}
-
-// static void	assign_array(t_env *head, char **array)
-// {
-// 	int		i;
-// 	t_env	*temp;
-
-// 	i = 0;
-// 	temp = head;
-// 	while (temp)
-// 	{
-// 		array[i] = (char *)malloc(strlen(temp->key) + strlen(temp->value) + 2);
-// 		sprintf(array[i], "%s=%s", temp->key, temp->value);
-// 		temp = temp->next;
-// 		i++;
-// 	}
-// 	array[i] = NULL;
-// }
-
-// char	**env_to_char_array(t_env *head)
-// {
-// 	int		count;
-// 	char	**result;
-
-// 	count = count_env_variables(head);
-// 	result = (char **)malloc((count + 1) * sizeof(char *));
-// 	assign_array(head, result);
-// 	return (result);
-// }
-char	*find_path(t_cmdtable *table, char **env, bool last_cmd)
-{
-	char		*temp;
-	struct stat	path_stat;
-
-	temp = NULL;
-	(void) last_cmd;
-	if (stat(table->cmd[0], &path_stat) == 0)
-	{
-		if (S_ISDIR(path_stat.st_mode))
-		{
-			ft_putstr_fd(table->cmd[0], STDERR_FILENO);
-			ft_putstr_fd(" is a directory\n", STDERR_FILENO);
-			temp = NULL;
-		}
-	}
-	if (access(table->cmd[0], X_OK) == 0)
-		temp = table->cmd[0];
-	else
-		temp = cmd_path(table->cmd[0], env);
-	if (!temp)
-	{
-		if (last_cmd)
-			dup2(table->fd_in, STDIN_FILENO);
-		exit(127);
-	}
-	return (temp);
 }
 
 int	update_env_value(t_env *head, const char *key, const char *new_value)
